Changement Administration + Modification local
- Ajout de la fonctionnalité pour un administrateur de modifier un local
- Export de toutes les fonctions d'Administration dans un fichier afin de simplifier la lecture du fichier Administration.tsx
- Changement de nom de Administration.js en AdministrationRoom.js ce qui permettre de créer, plus tard, un fichier AdministrationImplantation.js pour avoir les fonctions dans 2 fichiers séparés => facilite la lecture
